Apply code-review findings: defence-in-depth, test de-duplication, builder ergonomics, module split
- Add defence-in-depth comment on the redundant entrypoint-is_relative check
  in validate_location_matches_manifest (LoadedSkillLocation::new already
  enforces this at construction).

- Extract make_mismatched_parts helper and parametrize the two mismatched-
  identifier tests into a single rstest with two cases covering both `new`
  and `set_location` rejection paths.

- Extend arb_relative_path Strategy to generate single-level nested paths
  (e.g. "subdir/SKILL.md") alongside flat filenames via prop_oneof.

- Add .root() setter to TestSkillBuilder so callers can set the runtime
  root without constructing a LoadedSkillLocation manually.  Simplify
  make_skill in selector.rs to use .root("/tmp/test").

- Split src/skills/mod.rs (was 852 lines) into:
  - escape.rs: escape_xml_attr, escape_skill_content, normalize_line_endings
  - tests.rs: all unit tests under #[cfg(test)]
  Re-export escape helpers for backward compatibility.
